Add client data
The `clients.json` file contains user's PII, such as full names and email addresses. Ideally, no PII data should be committed to git. However, assuming this is just sample data and not actual user information, it is acceptable for it to be committed.

Moving forward, we should consider alternatives like storing the data in an S3 bucket and reading from it, or keeping encrypted files in the database, rather than committing sensitive data to git.